bots/gsbot/requirements.txt
# Discord Bot
discord.py>=2.3.2
python-dotenv>=1.0.0

# Database
sqlalchemy>=2.0.23
alembic>=1.13.1

# Data Processing
pandas>=2.1.4
numpy>=1.26.2

# API and Web
aiohttp>=3.9.1
requests>=2.31.0

# Caching
cachetools>=5.3.2

# Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# Code Quality
black>=23.12.1
flake8>=6.1.0
mypy>=1.7.1
pylint>=3.0.3

# Utilities
pyyaml>=6.0.1
colorlog>=6.8.0
